requirements.txt
matplotlib==3.5.1
numpy==1.21.5
pandas==1.5.2
Pillow==9.3.0
scikit_learn==1.1.3
seaborn==0.12.1
streamlit==1.15.1
xgboost==1.7.1
